+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - `RUST_LOG=ab_av1=info` "info" level logs various progress results like sample encode info
   - `RUST_LOG=ab_av1=debug` "debug" level logs include ffmpeg calls
 * Don't panic on non-zero status exit.
+* When unable to parse a vmaf score fail faster and include ffmpeg output.
 
 # v0.7.14
 * Fix bash completions of some filenames.